change project type to DLL (never to be merged with master)
change project type to DLL to avoid that F-Secure considers stubby as a virus.

note that doing so we change the name of the nuget, the name of the assembly.

Ideally, we would have 2 projects (stubby and stubbydll) but this was too much changes for now ;)
In the future, this branch will be deleted and replaced with the better approach